Config.xlsx: Added the ProcessedData_File entry
Main.xaml : Added in the initialization the building of 2 data table (dt_SuccessData and dt_ExceptionData) which will contain the processed data information and its status.
Also added as input / output arguments in the success and exception data table when invoking the SetTransactionStatus workflow
Lastly, added the sequence for generating the processed data report.

SetTransactionStatus.xaml : Added the addition io arguments for the success and exception data table. Added the add data row activity for the success data in the success flow. Added the add data row activity for the exception data in the business exception flow.

PRocess.xaml - added an activity to throw a business exception whenever data being processed contains a blank username.